Fix frontend build and prepare backend for deployment
This commit addresses two main goals:
1.  Fixes the frontend build issues.
2.  Prepares the backend for deployment to Render.

**Frontend:**
- Refactored the collections page to correctly use client and server components with Suspense boundaries, fixing the `useSearchParams` error.
- Added a Suspense boundary to the root layout to fix issues with client-side providers.
- Added a dummy Stripe publishable key to a `.env.local` file to allow the build to pass.
- Added defensive code to the profile page to prevent errors when user info is not available during prerendering.

**Backend:**
- Created a `render.yaml` file to define the services for deployment on Render.
- Updated the `README.md` with instructions on how to deploy the backend to Render.
