+++ render.yaml
@@ -1,0 +1,60 @@
+# render.yaml
+services:
+  # Backend Service
+  - type: web
+    name: nike-shop-backend
+    env: node
+    repo: https://github.com/your-repo/nike-shop # Replace with your repo URL
+    plan: free # Or your desired plan
+    healthCheckPath: /
+    buildCommand: "cd backend && npm install && npm run build"
+    startCommand: "cd backend && npm run start"
+    envVars:
+      - key: NODE_ENV
+        value: production
+      - key: DATABASE_URL
+        fromDatabase:
+          name: nike-shop-db
+          property: connectionString
+      - key: API_URL
+        value: "https://your-backend-url.onrender.com" # Replace with your backend URL
+      - key: FRONTEND_URL
+        value: "https://your-frontend-url.onrender.com" # Replace with your frontend URL
+      - key: PORT
+        value: 10000
+      - key: VERSION
+        value: "v1"
+      - key: BETTER_AUTH_SECRET
+        sync: false
+      - key: GOOGLE_CLIENT_ID
+        sync: false
+      - key: GOOGLE_CLIENT_SECRET
+        sync: false
+      - key: GITHUB_CLIENT_ID
+        sync: false
+      - key: GITHUB_CLIENT_SECRET
+        sync: false
+      - key: STRIPE_SECRET_KEY
+        sync: false
+      - key: STRIPE_WEBHOOK_SECRET
+        sync: false
+
+  # PostgreSQL Database
+  - type: psql
+    name: nike-shop-db
+    plan: free # Or your desired plan
+    pgaas:
+      version: "14" # Or your desired PostgreSQL version
+      shadow: false
+      ipAllowList: []
+      maintenanceWindow:
+        day: "Sunday"
+        hour: 3
+        minute: 0
+        timezone: "UTC"
+      backups:
+        enabled: true
+        frequency: "daily"
+        retention: 7
+      highAvailability:
+        enabled: false
